feat: add productCard component for list all products
