+++ package.json
@@ -9,6 +9,9 @@
     "lint": "next lint",
     "prisma:generate": "prisma generate"
   },
+  "prisma": {
+    "seed": "node prisma/seed.js"
+  },
   "dependencies": {
     "@prisma/client": "^6.6.0",
     "next": "15.3.0",
@@ -18,6 +21,7 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@prisma/client": "^6.6.0",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
     "@types/react": "^19",
@@ -25,8 +29,7 @@
     "eslint": "^9",
     "eslint-config-next": "15.3.0",
     "tailwindcss": "^4",
-    "typescript": "^5",
-    "@prisma/client": "^6.6.0"
-
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
   }
 }
